feat: Make Google ADK backend the default route
- Replace Vercel AI SDK with Google ADK integration at root route (/)
- Remove authentication requirement for main chat interface
- Create guest session automatically for seamless access
- Backup original Vercel AI SDK implementation

This change makes the Google ADK backend-powered chat the primary
interface, eliminating the need for authentication and providing
direct access to the enhanced chat features with backend integration.
