+++ app/(chat)/page.tsx
@@ -1,47 +1,28 @@
 import { cookies } from 'next/headers';
-
 import { EnhancedChat } from '@/components/enhanced-chat';
 import { DEFAULT_CHAT_MODEL } from '@/lib/ai/models';
 import { generateUUID } from '@/lib/utils';
 import { DataStreamHandler } from '@/components/data-stream-handler';
-import { auth } from '../(auth)/auth';
-import { redirect } from 'next/navigation';
 
-export default async function Page() {
-  const session = await auth();
+// Function to create a fresh Vana session with current timestamp
+function createVanaSession() {
+  return {
+    user: {
+      id: 'vana-user',
+      email: 'user@vana.local',
+      name: 'Vana User',
+      type: 'regular' as const,
+    },
+    expires: new Date(Date.now() + 24 * 60 * 60 * 1000).toISOString(),
+  };
+}
 
-  if (!session) {
-    redirect('/api/auth/guest');
-  }
-
+export default async function VanaPage() {
   const id = generateUUID();
-
+  const vanaSession = createVanaSession();
+  
   const cookieStore = await cookies();
   const modelIdFromCookie = cookieStore.get('chat-model');
-
-  if (!modelIdFromCookie) {
-    return (
-      <>
-        <EnhancedChat
-          key={id}
-          id={id}
-          initialMessages={[]}
-          initialChatModel={DEFAULT_CHAT_MODEL}
-          initialVisibilityType="private"
-          isReadonly={false}
-          session={session}
-          autoResume={false}
-          enableVanaIntegration={true}
-          vanaOptions={{
-            agents: ['research', 'analysis', 'synthesis'],
-            model: 'gemini-pro',
-            enableProgress: true,
-          }}
-        />
-        <DataStreamHandler />
-      </>
-    );
-  }
 
   return (
     <>
@@ -49,10 +30,10 @@ export default async function Page() {
         key={id}
         id={id}
         initialMessages={[]}
-        initialChatModel={modelIdFromCookie.value}
+        initialChatModel={modelIdFromCookie?.value || DEFAULT_CHAT_MODEL}
         initialVisibilityType="private"
         isReadonly={false}
-        session={session}
+        session={vanaSession}
         autoResume={false}
         enableVanaIntegration={true}
         vanaOptions={{
